Fix cluster.rc tests with a hack for now (#4508)
On the latest version of ubuntu all tests that have cluster.rc are
failing. Mostly because of this:
https://datatracker.ietf.org/doc/html/rfc6890

The following patch fixes it for now. But we may need to change the
framework to fix it properly I think. Ideally localhost or loopback ip
range shouldn't be allowed as hostname. This part we may have to address
separately.

Fixes: #4457
Change-Id: Iee3e6710c301cde4d5bb250ff45fd84031a878be

Signed-off-by: Pranith Kumar Karampuri <pranith.karampuri@phonepe.com>