api/glfs: Fix use after free in readdir (#4477) (#4581)
There is a code path in glfs readdir where we might end up
using a free'ed memory. A possible case is when we use the
fd to read the offset 0 after reading the content at least
once, and the during the second time, the readdir return 0,
probably due to the contents got deleted.

  > cherry-picked from commit 490e018484
Change-Id: I08fa6cc9b924b90597ef4498e081cae9339fbd6a
Fixes: #3955

Signed-off-by: Mohammed Rafi KC <rafi.kavungal@iternity.com>